universal-agent/requirements.txt
# Core Flask and WebSocket
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0

# Authentication
PyJWT==2.8.0
python-jose==3.3.0
requests==2.31.0
cryptography==41.0.7
jwcrypto==1.5.6

# Device Communication
spidev==3.6
gpiod==2.1.0
pyserial==3.5
pyusb==1.3.1
hidapi==0.14.0
# bluetooth-tools==0.2.2  # Removed - package doesn't exist on PyPI
# Bluetooth for EMOTIV devices
# pybluez2==0.46  # Removed - requires bluetooth development headers
# bluepy==1.3.0  # Removed - requires glib-2.0-dev which may not be available
bleak==0.21.1

# Signal Processing  
numpy==1.24.3
scipy==1.11.4

# Data Handling
pandas==2.0.3
python-dotenv==1.0.0

# Encryption
pycryptodome==3.19.0

# System and Hardware
RPi.GPIO==0.7.1
psutil==5.9.6

# Async and Threading
asyncio-mqtt==0.16.1

# Logging and Monitoring
structlog==23.2.0
colorlog==6.8.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0